sdks/queue-az/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "queue-az"
version = "0.1.0"
authors = [
    {name="Marcel Claramunt", email="marcel@moveread.com"}
]
description = "Async Python SDK for Azure Queue Storage"
dependencies = [
    "azure-storage-queue", "aiohttp"
]
requires-python = ">=3.10"
readme = {file="README.md", content-type="text/markdown"}

[project.urls]
home = "https://github.com/moveread/azure-sdks/tree/main/sdks/queue-az"
repo = "https://github.com/moveread/azure-sdks"


# [project.optional-dependencies]
# test = [
#   "pytest < 5.0.0",
#   "pytest-cov[all]"
# ]